aura_chain/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
anthropic==0.18.1
google-generativeai==0.3.2
openai==1.12.0
pandas==2.2.0
numpy==1.26.3
redis==5.0.1
sqlalchemy==2.0.25
asyncpg==0.29.0
httpx==0.26.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
loguru==0.7.2
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
aiofiles==23.2.1
plotly==5.18.0
scipy==1.12.0
scikit-learn==1.4.0
joblib==1.3.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.26.0
prophet>=1.1.5
cmdstanpy>=1.2.0
holidays>=0.40.0
plotly>=5.18.0
pytrends==4.9.2
feedparser==6.0.10
beautifulsoup4==4.12.2